capstone_sys::cs_free

pub unsafe extern fn cs_free(insn: *mut cs_insn, count: size_t)

Free a Capstone allocated array of instruction.

Safety

insn shall originate either from a previous call to cs_malloc, in which case the count should be 1, or cs_disasm in which case the count should be the return value of cs_disasm