Function capstone_sys::cs_regs_access

pub unsafe extern "C" fn cs_regs_access(
    handle: csh,
    insn: *const cs_insn,
    regs_read: *mut u16,
    regs_read_count: *mut u8,
    regs_write: *mut u16,
    regs_write_count: *mut u8
) -> Type

Retrieve all the registers accessed by an instruction, either explicitly or implicitly.

WARN: when in ‘diet’ mode, this API is irrelevant because engine does not store registers.

@handle: handle returned by cs_open() @insn: disassembled instruction structure returned from cs_disasm() or cs_disasm_iter() @regs_read: on return, this array contains all registers read by instruction. @regs_read_count: number of registers kept inside @regs_read array. @regs_write: on return, this array contains all registers written by instruction. @regs_write_count: number of registers kept inside @regs_write array.

@return CS_ERR_OK on success, or other value on failure (refer to cs_err enum for detailed error).